packaging: rpm: don't use cpack release version as python optimization
The current package relase number (which happens to be 2) is for reasons
unknown used as optimization level when installing python modules.

Also, setuptools generates files with optimization level 0, 1, and 2,
but only lists level 0 and 2 in its file record, making rpm complain
about unpackaged files.

Fix both of the above by setting the optimization level to 1.

Signed-off-by: Martin Hundebøll <mhu@silicom.dk>